refac: Settlement report job structure
