+++ source/__init__.py
@@ -11,15 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import uuid
-from datetime import datetime
-import configargparse
-
-
-def valid_date(s: str) -> datetime:
-    """See https://stackoverflow.com/questions/25470844/specify-date-format-for-python-argparse-input-arguments"""
-    try:
-        return datetime.strptime(s, "%Y-%m-%dT%H:%M:%SZ")
-    except ValueError:
-        msg = "not a valid date: {0!r}".format(s)
-        raise configargparse.ArgumentTypeError(msg)
