+++ source/databricks/__init__.py
@@ -11,15 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from pyspark.sql import DataFrame
-from pyspark.sql.session import SparkSession
-
-from settlement_report_job.settlement_report_args import SettlementReportArgs
-
-
-def create_time_series(
-    spark: SparkSession,
-    args: SettlementReportArgs,
-) -> DataFrame:
-    # ToDo JMG: implement
-    return spark.createDataFrame([], schema=[])
