+++ source/databricks/settlement_report/settlement_report_job/domain/settlement_report_args.py
@@ -15,8 +15,8 @@ from uuid import UUID
 from dataclasses import dataclass
 from datetime import datetime
 
-from settlement_report_job.calculation_type import CalculationType
-from settlement_report_job.market_role import MarketRole
+from settlement_report_job.domain.calculation_type import CalculationType
+from settlement_report_job.domain.market_role import MarketRole
 
 
 @dataclass
