+++ source/databricks/settlement_report/settlement_report_job/domain/time_series_factory.py
@@ -11,24 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import uuid
-from datetime import datetime
+from pyspark.sql import DataFrame
+from pyspark.sql.session import SparkSession
 
-import pytest
-
-from settlement_report_job.domain.calculation_type import CalculationType
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
 
 
-@pytest.fixture(scope="session")
-def any_settlement_report_args() -> SettlementReportArgs:
-    return SettlementReportArgs(
-        report_id=str(uuid.uuid4()),
-        period_start=datetime(2018, 3, 31, 22, 0, 0),
-        period_end=datetime(2018, 4, 30, 22, 0, 0),
-        calculation_type=CalculationType.WHOLESALE_FIXING,
-        split_report_by_grid_area=True,
-        prevent_large_text_files=False,
-        time_zone="Europe/Copenhagen",
-        catalog_name="catalog_name",
-    )
+def create_time_series(
+    spark: SparkSession,
+    args: SettlementReportArgs,
+) -> DataFrame:
+    # ToDo JMG: implement
+    return spark.createDataFrame([], schema=[])
