+++ source/databricks/settlement_report/settlement_report_job/infrastructure/settlement_report_job_args.py
@@ -19,13 +19,13 @@ from argparse import Namespace
 import configargparse
 from configargparse import argparse
 
-from settlement_report_job import logging_configuration
-from settlement_report_job.args_helper import valid_date
-from settlement_report_job.calculation_type import CalculationType
+from settlement_report_job.infrastructure import logging_configuration
+from settlement_report_job.infrastructure.args_helper import valid_date
+from settlement_report_job.domain.calculation_type import CalculationType
 from settlement_report_job.logger import Logger
-from settlement_report_job.market_role import MarketRole
-from settlement_report_job.settlement_report_args import SettlementReportArgs
-import settlement_report_job.environment_variables as env_vars
+from settlement_report_job.domain.market_role import MarketRole
+from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
+import settlement_report_job.infrastructure.environment_variables as env_vars
 
 
 def parse_command_line_arguments() -> Namespace:
